feat: update default grid column order
Update alter_field.order to patch the default grid-view column metadata
so NocoDB's table UI reflects migration-defined field order.

Resolves #12

+++ internal/migration/parser_test.go
@@ -123,7 +123,7 @@ func TestValidateMigration(t *testing.T) {
 			name: "alter_field with order",
 			migration: &Migration{
 				Operations: []Operation{
-					{Type: "alter_field", Table: "T", Column: &ColumnDefinition{Name: "c", Type: "Number", Order: func() *int { v := 3; return &v }()}},
+					{Type: "alter_field", Table: "T", Column: &ColumnDefinition{Name: "c", Type: "Number", Order: func() *float64 { v := 3.0; return &v }()}},
 				},
 			},
 			wantError: false,
